+++ Plugins/Player/Player.hpp
@@ -56,6 +56,8 @@ private:
     ArgumentStack ToggleDM                          (ArgumentStack&& args);
     ArgumentStack SetObjectMouseCursorOverride      (ArgumentStack&& args);
     ArgumentStack SetObjectHiliteColorOverride      (ArgumentStack&& args);
+    ArgumentStack RemoveEffectFromTURD              (ArgumentStack&& args);
+    ArgumentStack SetSpawnLocation                  (ArgumentStack&& args);
 
     CNWSPlayer *player(ArgumentStack& args);